chore: move the scipy adapter test to appropriate dir location
